+++ packages/contracts/src/plugins/counter-example/v2/CounterV2.sol
@@ -1,4 +1,4 @@
-// SPDX-License-Identifier: MIT
+// SPDX-License-Identifier: AGPL-3.0-or-later
 
 pragma solidity 0.8.17;
 
@@ -22,11 +22,6 @@ contract CounterV2 is PluginUUPSUpgradeable {
     /// @notice A new varaible added in V2.
     /// @dev By appending a new variable, the existing storage gets modified.
     uint256 public newVariable;
-
-    /// @dev Used to disallow initializing the implementation contract by an attacker for extra safety.
-    constructor() {
-        _disableInitializers();
-    }
 
     /// @notice Initializes the plugin.
     /// @param _dao The contract of the associated DAO.