+++ packages/contracts/CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Set the implementation address in the constructor of `PluginSetup`.
+- Moved `_disableInitializers` inside the constructors of `PluginCloneable` and `PluginUUPSUpgradeable`.
+- Changed license from `MIT` to `AGPL-3.0-or-later`.
 - Renamed `DAOSettings.name` to `DAOSettings.subdomain`.
 - Fixed the ERC165 interface ID and `supportsInterface` computations.
 - Replaced non-upgradeable contract dependencies by upgradeable ones for the cloneable `Admin` plugin.